requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
biopython==1.85
fastapi==0.115.14
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.73.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.4
lightning-utilities==0.14.3
Markdown==3.8.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.3
numpy==2.3.1
packaging==24.1
pandas==2.3.0
pandas-stubs==2.2.3.250527
pathspec==0.12.1
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
ruff==0.12.1
seaborn
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1
torchmetrics==1.7.3
tqdm==4.67.1
types-pytz==2025.2.0.20250516
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
yarl==1.20.1
onnx
onnxruntime
httpx
transformers
pre-commit==3.7.1
pytest-cov==5.0.0
